Move input util functions to input mod
